+++ packages/playground/src/components/Editors.tsx
@@ -1,0 +1,147 @@
+import { useCallback, useState } from 'react';
+import MonacoEditor from '@monaco-editor/react';
+import isEqualWith from 'lodash/isEqualWith';
+
+import type { ErrorSchema, RJSFSchema, UiSchema } from '@rjsf/utils';
+
+const monacoEditorOptions = {
+  minimap: {
+    enabled: false,
+  },
+  automaticLayout: true,
+};
+
+interface EditorProps {
+  title: string;
+  code: string;
+  onChange: (code: string) => void;
+}
+
+function Editor({ title, code, onChange }: EditorProps) {
+  const [valid, setValid] = useState(true);
+
+  const onCodeChange = useCallback(
+    (code: string | undefined) => {
+      if (!code) {
+        return;
+      }
+
+      try {
+        const parsedCode = JSON.parse(code);
+        setValid(true);
+        onChange(parsedCode);
+      } catch (err) {
+        setValid(false);
+      }
+    },
+    [setValid, onChange]
+  );
+
+  const icon = valid ? 'ok' : 'remove';
+  const cls = valid ? 'valid' : 'invalid';
+
+  return (
+    <div className='panel panel-default'>
+      <div className='panel-heading'>
+        <span className={`${cls} glyphicon glyphicon-${icon}`} />
+        {' ' + title}
+      </div>
+      <MonacoEditor
+        language='json'
+        value={code}
+        theme='vs-light'
+        onChange={onCodeChange}
+        height={400}
+        options={monacoEditorOptions}
+      />
+    </div>
+  );
+}
+
+const toJson = (val: unknown) => JSON.stringify(val, null, 2);
+
+type EditorsProps = {
+  schema: RJSFSchema;
+  setSchema: React.Dispatch<React.SetStateAction<RJSFSchema>>;
+  uiSchema: UiSchema;
+  setUiSchema: React.Dispatch<React.SetStateAction<UiSchema>>;
+  formData: any;
+  setFormData: React.Dispatch<React.SetStateAction<any>>;
+  extraErrors: ErrorSchema | undefined;
+  setExtraErrors: React.Dispatch<React.SetStateAction<ErrorSchema | undefined>>;
+  setShareURL: React.Dispatch<React.SetStateAction<string | null>>;
+};
+
+export default function Editors({
+  extraErrors,
+  formData,
+  schema,
+  uiSchema,
+  setExtraErrors,
+  setFormData,
+  setSchema,
+  setShareURL,
+  setUiSchema,
+}: EditorsProps) {
+  const onSchemaEdited = useCallback(
+    (newSchema) => {
+      setSchema(newSchema);
+      setShareURL(null);
+    },
+    [setSchema, setShareURL]
+  );
+
+  const onUISchemaEdited = useCallback(
+    (newUiSchema) => {
+      setUiSchema(newUiSchema);
+      setShareURL(null);
+    },
+    [setUiSchema, setShareURL]
+  );
+
+  const onFormDataEdited = useCallback(
+    (newFormData) => {
+      if (
+        !isEqualWith(newFormData, formData, (newValue, oldValue) => {
+          // Since this is coming from the editor which uses JSON.stringify to trim undefined values compare the values
+          // using JSON.stringify to see if the trimmed formData is the same as the untrimmed state
+          // Sometimes passing the trimmed value back into the Form causes the defaults to be improperly assigned
+          return JSON.stringify(oldValue) === JSON.stringify(newValue);
+        })
+      ) {
+        setFormData(newFormData);
+        setShareURL(null);
+      }
+    },
+    [formData, setFormData, setShareURL]
+  );
+
+  const onExtraErrorsEdited = useCallback(
+    (newExtraErrors) => {
+      setExtraErrors(newExtraErrors);
+      setShareURL(null);
+    },
+    [setExtraErrors, setShareURL]
+  );
+
+  return (
+    <div className='col-sm-7'>
+      <Editor title='JSONSchema' code={toJson(schema)} onChange={onSchemaEdited} />
+      <div className='row'>
+        <div className='col-sm-6'>
+          <Editor title='UISchema' code={toJson(uiSchema)} onChange={onUISchemaEdited} />
+        </div>
+        <div className='col-sm-6'>
+          <Editor title='formData' code={toJson(formData)} onChange={onFormDataEdited} />
+        </div>
+      </div>
+      {extraErrors && (
+        <div className='row'>
+          <div className='col'>
+            <Editor title='extraErrors' code={toJson(extraErrors || {})} onChange={onExtraErrorsEdited} />
+          </div>
+        </div>
+      )}
+    </div>
+  );
+}
