+++ packages/playground/src/components/Playground.tsx
@@ -1,6 +1,5 @@
 import { useCallback, useState, useRef, useEffect, ComponentType, FormEvent } from 'react';
 import 'react-app-polyfill/ie11';
-import isEqualWith from 'lodash/isEqualWith';
 
 import { withTheme, IChangeEvent, FormProps } from '@rjsf/core';
 import {
@@ -19,96 +18,9 @@ import DemoFrame from './DemoFrame';
 import ErrorBoundary from './ErrorBoundary';
 import GeoPosition from './GeoPosition';
 import { ThemesType } from './ThemeSelector';
-import Editor from './Editor';
+import Editors from './Editors';
 
 const log = (type: string) => console.log.bind(console, type);
-const toJson = (val: unknown) => JSON.stringify(val, null, 2);
-
-type EditorsProps = {
-  schema: RJSFSchema;
-  setSchema: React.Dispatch<React.SetStateAction<RJSFSchema>>;
-  uiSchema: UiSchema;
-  setUiSchema: React.Dispatch<React.SetStateAction<UiSchema>>;
-  formData: any;
-  setFormData: React.Dispatch<React.SetStateAction<any>>;
-  extraErrors: ErrorSchema | undefined;
-  setExtraErrors: React.Dispatch<React.SetStateAction<ErrorSchema | undefined>>;
-  setShareURL: React.Dispatch<React.SetStateAction<string | null>>;
-};
-
-function Editors({
-  extraErrors,
-  formData,
-  schema,
-  uiSchema,
-  setExtraErrors,
-  setFormData,
-  setSchema,
-  setShareURL,
-  setUiSchema,
-}: EditorsProps) {
-  const onSchemaEdited = useCallback(
-    (newSchema) => {
-      setSchema(newSchema);
-      setShareURL(null);
-    },
-    [setSchema, setShareURL]
-  );
-
-  const onUISchemaEdited = useCallback(
-    (newUiSchema) => {
-      setUiSchema(newUiSchema);
-      setShareURL(null);
-    },
-    [setUiSchema, setShareURL]
-  );
-
-  const onFormDataEdited = useCallback(
-    (newFormData) => {
-      if (
-        !isEqualWith(newFormData, formData, (newValue, oldValue) => {
-          // Since this is coming from the editor which uses JSON.stringify to trim undefined values compare the values
-          // using JSON.stringify to see if the trimmed formData is the same as the untrimmed state
-          // Sometimes passing the trimmed value back into the Form causes the defaults to be improperly assigned
-          return JSON.stringify(oldValue) === JSON.stringify(newValue);
-        })
-      ) {
-        setFormData(newFormData);
-        setShareURL(null);
-      }
-    },
-    [formData, setFormData, setShareURL]
-  );
-
-  const onExtraErrorsEdited = useCallback(
-    (newExtraErrors) => {
-      setExtraErrors(newExtraErrors);
-      setShareURL(null);
-    },
-    [setExtraErrors, setShareURL]
-  );
-
-  return (
-    <div className='col-sm-7'>
-      <Editor title='JSONSchema' code={toJson(schema)} onChange={onSchemaEdited} />
-      <div className='row'>
-        <div className='col-sm-6'>
-          <Editor title='UISchema' code={toJson(uiSchema)} onChange={onUISchemaEdited} />
-        </div>
-        <div className='col-sm-6'>
-          <Editor title='formData' code={toJson(formData)} onChange={onFormDataEdited} />
-        </div>
-      </div>
-      {extraErrors && (
-        <div className='row'>
-          <div className='col'>
-            <Editor title='extraErrors' code={toJson(extraErrors || {})} onChange={onExtraErrorsEdited} />
-          </div>
-        </div>
-      )}
-    </div>
-  );
-}
 
 export interface LiveSettings {
   showErrorList: false | 'top' | 'bottom';
